+++ omicverse_guide/docs/Release_notes.md
@@ -438,6 +438,8 @@ Support Raw Windows platform
 ### Space Module
 
 - Fixed an error in `construct_landscape` of `ov.space.STT`
+- Fixed an error of `get_image_idx_1D` in `ov.space.svg` #117
+- Added `COMMOT` to calculate the cell-cell interaction of spatial RNA-seq.
 
 ### Pl Module